refactor: Unify the two runGrammarChecker functions
This also changes the behavior where it would
refresh the DOM on the first results received from the
backend but now it will wait for all of them to be received
and then save some render cycles or CPU if the browser
is slow and the users start clicking on the results while
others are being rendered.

This also improves some other small things
which should be such a big hit on performance but
are still nice to be done (e.g. cloning arrays outside
of a for loop :))